refactor: ShellNavbar paths
