+++ app/ui/src/app/components/ShellNavbar.tsx
